requirements.txt
streamlit
requests
google-adk
google-generativeai
python-dotenv
fastapi
uvicorn